PDFBOX-5660: remove redundant cast

+++ src/main/java/org/apache/pdfbox/jbig2/segments/TextRegion.java
@@ -249,7 +249,7 @@ public class TextRegion implements Region
         amountOfSymbolInstances = subInputStream.readBits(32) & 0xffffffff;
 
         // sanity check: don't decode more than one symbol per pixel
-        long pixels = (long) regionInfo.getBitmapWidth() * (long) regionInfo.getBitmapHeight();
+        long pixels = regionInfo.getBitmapWidth() * (long) regionInfo.getBitmapHeight();
         if (pixels < amountOfSymbolInstances)
         {
             amountOfSymbolInstances = pixels;
